webapi/webapi-webspeech-w3c-tests/webspeech/blink/speech-synthesis-speak-invalid-argument-throws.html

<!--
Test convert from Blink:
https://github.com/crosswalk-project/blink-crosswalk/blob/master/LayoutTests/fast/speechsynthesis/speech-synthesis-speak-invalid-argument-throws.html
-->

<!DOCTYPE html>
<meta charset="utf-8">
<title>Web Speech API Test: passing an argument other than a SpeechSynthesisUtterance to speechSynthesis.speak throws a TypeError exception</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://dvcs.w3.org/hg/speech-api/raw-file/tip/speechapi.html">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

test(function () {
  var num = 0;
  try{
    speechSynthesis.speak();
  } catch (e) {
    assert_equals(e.message, "Failed to execute \'speak\' on \'SpeechSynthesis\': 1 argument required, but only 0 present.");
    num++;
  }
  try{
    speechSynthesis.speak(0);
  } catch (e) {
    assert_equals(e.message, "Failed to execute \'speak\' on \'SpeechSynthesis\': Invalid utterance argument");
    num++;
  }
  try{
    speechSynthesis.speak('');
  } catch (e) {
    assert_equals(e.message, "Failed to execute \'speak\' on \'SpeechSynthesis\': Invalid utterance argument");
    num++;
  }
  try{
    speechSynthesis.speak(document.body);
  } catch (e) {
    assert_equals(e.message, "Failed to execute \'speak\' on \'SpeechSynthesis\': Invalid utterance argument");
    num++;
  }
  try{
    speechSynthesis.speak({});
  } catch (e) {
    assert_equals(e.message, "Failed to execute \'speak\' on \'SpeechSynthesis\': Invalid utterance argument");
    num++;
  }
  try{
    speechSynthesis.speak((new SpeechSynthesisUtterance()).text = 'hello');
  } catch (e) {
    assert_equals(e.message, "Failed to execute \'speak\' on \'SpeechSynthesis\': Invalid utterance argument");
    num++;
  }
  assert_equals(num, 6, "the number of exception");
}, "Test that passing an argument other than a SpeechSynthesisUtterance to speechSynthesis.speak throws a TypeError exception");

</script>
